Add system to enable sprite layers as markers
MarkerSystem only allows showing and hiding entire sprites, which we
don't want in this case. All other pre-existing systems lose out to the
T-ray scanner, which always shows all layers. We want to make some
sprite layers only visible when markers are shown in the sandbox, so we
have to add a new system that dynamically adds/removes sprite layers
depending on whether or not the markers are supposed to be visible.